refactor: move mocket.entry to mocket.compat.entry
