+++ mocket/__init__.py
@@ -1,15 +1,13 @@
-from mocket.compat import FakeSSLContext
+from mocket.compat import FakeSSLContext, MocketEntry
 from mocket.core.async_mocket import async_mocketize
 from mocket.core.mocket import Mocket
 from mocket.core.mocketizer import Mocketizer, mocketize
 from mocket.core.socket import MocketSocket
 from mocket.core.ssl.context import MocketSSLContext
 from mocket.core.ssl.socket import MocketSSLSocket
-from mocket.entry import MocketEntry
 
 __all__ = [
     "Mocket",
-    "MocketEntry",
     "MocketSSLContext",
     "MocketSSLSocket",
     "MocketSocket",
@@ -18,6 +16,7 @@ __all__ = [
     "mocketize",
     # NOTE this is here for backwards-compat to keep old import-paths working
     "FakeSSLContext",
+    "MocketEntry",
 ]
 
 __version__ = "3.13.2"
