+++ mocket/compat/__init__.py
@@ -1,5 +1,7 @@
+from mocket.compat.entry import MocketEntry
 from mocket.core.ssl.context import MocketSSLContext as FakeSSLContext
 
 __all__ = [
     "FakeSSLContext",
+    "MocketEntry",
 ]
