+++ mocket/compat/entry.py
